requirements.txt
google-generativeai 
requests 
loguru 
yfinance 
fastapi 
uvicorn 
python-dotenv